+++ mobile/src/domain/repositories/user-repository.ts
@@ -6,4 +6,5 @@ export interface DashboardData {
 
 export interface UserRepository {
     getDashboardData(role: string): Promise<DashboardData>;
+    createUser(name: string, email: string): Promise<any>;
 }
